+++ Metalord_btin/MetaLord/Assets/MainLevel/LowPolyFoodLite/LowPolyFoodLite_2020.3.0_HDRP_v1.0.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e8d1d59b4af53fb4685894e0972082fc
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
