+++ Metalord_btin/MetaLord/Assets/MainLevel/LowPolyInterior/LowPolyInterior_2020.3_HDRP_v1.42.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 928786dcd45751645a936e7b70e3d3c4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
