+++ Metalord_btin/MetaLord/Assets/MainLevel/LowPolyInterior/LowPolyInterior_2020.3_URP_v1.42.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d647bfaf5780adb41a2b1db46e02a14d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
