+++ Metalord_btin/MetaLord/Assets/MainLevel/LowPolyMedievalPropsLite/LowPolyMedievalPropsLite_2020.3.0_HDRP_v1.0.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 518d1c1a2372e38479ca8a25bc9b0ee8
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
