+++ Metalord_btin/MetaLord/Assets/MainLevel/LowPolyMedievalPropsLite/Models/Bag_01.fbx.meta
@@ -48,7 +48,7 @@ ModelImporter:
     nodeNameCollisionStrategy: 0
     fileIdsGeneration: 2
     swapUVChannels: 0
-    generateSecondaryUV: 0
+    generateSecondaryUV: 1
     useFileUnits: 1
     keepQuads: 0
     weldVertices: 0
