+++ Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/Legacy/SSC_GunBase.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class SSC_GunBase : MonoBehaviour
+{
+
+}
