+++ Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/Legacy/SSC_GunBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1d2be578fc7f99646a8b43c476d97e3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
