--- Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/SSC_BondGun.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 72356db50a30c7242ab4c76f5a5afabb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
