--- Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/SSC_GrabGun.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 638c2489616e64440a77118a41528e87
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
