--- Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/SSC_GunBase.cs
@@ -1,6 +1,0 @@
-using UnityEngine;
-
-public class SSC_GunBase : MonoBehaviour
-{
-
-}
