--- Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/SSC_GunRot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f90190406a472364199640c96fcc08a5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
