--- Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/SSC_GunState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9155b1d1f61b117448a0f44fd105efd0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
