--- Metalord_btin/MetaLord/Assets/_Test/SSC/Scripts/SSC_PaintGun.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9f6912241efa03d4db8b9e29a6b69c4f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
